+++ CHANGELOG.md
@@ -1,1 +1,4 @@
 # 1.2.34
+- Fixed hammer giving infinite fuel
+- Fixed crash with hammer
+- Buffed hammer